==> code/01-gather-data-from-fortune-website.ipynb <==
# Gather Data from Fortune's Website

As the first step in the larger process of analyzing the press release language for Fortune 100 companies, I first need the list of Fortune 100 companies to work with. The purpose of this notebook is to do just that - gather the names, ranks and website urls for the companies in the Fortune 100. 

## Imports

import pandas as pd

from tqdm import tqdm
import time

from selenium import webdriver
from selenium.webdriver.common.by import By

import warnings
warnings.filterwarnings('ignore')

## Gather Fortune 100 list info

When I first started the data collection process, I attempted to use the `requests` and `bs4` libraries to gather information from the most recent [Fortune 100 list](https://fortune.com/fortune500/2020/search/). However, I quickly discovered that the page uses JavaScript, which the `requests` library was not able to collect. To overcome this, I found [`Selenium`](https://www.selenium.dev) and was able to use it to overcome the JavaScript barrier. Although `Selenium` is primarily used for automating and testing websites and web applications, I found it to be an incredibly powerful and useful library to work around some of the limitations of `requests`.

The code below uses the `By.XPATH` option and the `.click()` method to set the page to 100 Rows from the beginning. It then finds the top of the table that has all of the companies and then gathers all of the company names and ranks, as well as the link Fortune's profile page for each company (e.g., [Walmart's profile page](https://fortune.com/company/walmart/fortune500/). For each row in the table, it appends a dictionary to the `cos` list created earlier in the code, which is then used to create a DataFrame.

# prepare the option for the chrome driver
options = webdriver.ChromeOptions()
options.add_argument('headless')

# start chrome browser
browser = webdriver.Chrome(options=options)

# get all the data from the Fortune 100 list
browser.get('https://fortune.com/fortune500/2020/search/')

# create list that all of the info can be appended to
cos = []

# set the page length to 100 rows instead of 10 rows
options_rows = browser.find_elements(
    By.XPATH, '//select[@aria-label="rows per page"]/option')
for o in options_rows:
    if o.text == '100 Rows':
        o.click()

# gather all of the row data
elements = browser.find_elements(By.CLASS_NAME, 'rt-tr-group')

# iterate through the elements vairable to get the information that will
# be appended to the list `cos`
for e in tqdm(elements):
    # create `co` dictionary
    co = {}

    # get the company name, rank and link to the fortune profile page for each row
    co['company'] = e.text.split('\n')[1]
    co['rank'] = e.text.split('\n')[0]
    co['fortune_link'] = e.find_element(By.TAG_NAME, 'a').get_attribute('href')

    # append to the `cos` list
    cos.append(co)

    # sleep for 4 seconds so as not to hit the page too many times
    time.sleep(4)

# create a dataframe from the final list
cos = pd.DataFrame(cos)
cos

## Gather company website information from each Fortune profile page

Once I had gathered the names, ranks and profile page links for each company in the Fortune 100, I needed to get the link to the company's own website, and not just the link for the Fortune company profile. Given my prior experience with Fortune's website, I used `Selenium` again to gather the information for me and save it into the DataFrame in a new column, co_website. 

The code below iterates through each row in the DataFrame and uses the link collected from the table to visit each profile page for the Fortune 100, collect the link Fortune has listed as the company's website and add it to the DataFrame.

# prepare the option for the chrome driver
options = webdriver.ChromeOptions()
options.add_argument('headless')

# start chrome browser
browser = webdriver.Chrome(options=options)

# iterate through the links in the dataframe to get to the fortune
# profile page for each of the companies
for i in tqdm(range(len(cos))):
    url = cos.loc[i,'fortune_link']
    browser.get(f'{url}')
    vals = browser.find_elements(By.TAG_NAME, 'a')

    # get the company website from the fortune profile page
    cos.loc[i,'co_website'] = [
        v.get_attribute('href') for v in vals
        if 'fortune.com' not in v.get_attribute('href')
    ][2]

    time.sleep(3)

cos.head()

Following a review of the company websites in the co_website column, I found there were five website urls that Fortune's company profile listed and that were ultimately incorrect. To fix this, I've hard-coded the correct websites in the cell below to make sure they are corrected before the DataFrame is saved to the `./data/` directory.

cos.loc[0,'co_website'] = 'https://corporate.walmart.com'
cos.loc[13, 'co_website'] = 'https://investor.costco.com'
cos.loc[40,'co_website'] = 'http://www.freddiemac.com/investors/'
cos.loc[70, 'co_website'] = 'https://corporate.charter.com'
cos.loc[74, 'co_website'] = 'https://corporate.bestbuy.com'

# save the dataframe to a csv

cos.to_csv('../data/fortune_100_data.csv', index = False)
