==> code/05-gather-press-release-text.ipynb <==
# Gather Press Release Text

Now that I finally the links I need, I can start gathering the full text for the press releases. This notebook, while the shortest with just three code blocks, takes the longest to run. That's because we're gathering all 2,000+ press releases that we will use for our models. 

The code below iterates through each file, and then through each row in each file, to get the url to each specific press release. I've used the `requests` library instead of the `Selenium` library in this case because it both works and keeps the code a bit simpler than it otherwise would have to be. 

After using the `requests` library to call the url, the code runs the page contents through `BeautifulSoup` and collects the page's title and the body of the text. I also created a column for the full HTML as well just in case I needed it later. All of this data was then stored into a csv file saved in the `press_releases` folder to be cleaned and then used to model later on.

*Note: Similar to the notebook above, I have an extra column added in for Amazon so that I can use it to add the label in the next notebook.*

## Imports

import pandas as pd

import time
from tqdm import tqdm

import os

import requests
from bs4 import BeautifulSoup

import warnings
warnings.filterwarnings('ignore')

files = [f for f in os.listdir('../data/links/') if 'csv' in f]
files

for file in files:
    # create new file name that can be used later
    new_file_name = file.replace('links.csv', 'press_releases.csv')

    # read in the file as a data frame
    df = pd.read_csv(f'../data/links/{file}')

    # create list that dictionaries (created in for loop) can be appended to
    press_releases = []

    try:
        # iterate through each row in the data frame
        for i in tqdm(range(len(df))):
            time.sleep(3)
            try:
                
                # create dictionary for each row and the results it returns
                press_release = {}

                # for those files that have a base string, get the url, otherwise
                # just use the link column
                if type(df.loc[i, 'base']) == str:
                    url = df.loc[i, 'base'] + df.loc[i, 'link']
                else:
                    url = df.loc[i, 'link']

                req = requests.get(url)

                soup = BeautifulSoup(req.content, 'lxml')

                press_release['full_link'] = url
                press_release['title'] = soup.title.text
                press_release['body'] = soup.body.text
                press_release['html'] = soup
                
                if 'amazon' in file:
                    press_release['year'] = df.loc[i,'year']
                else:
                    pass
                    
                press_releases.append(press_release)
                

            except:
                print(f'Error: {file} | {url} | {i} | {req}')
            
    except:
        print(f'Error: {file} | {url}')

    pr_df = pd.DataFrame(press_releases)
    pr_df.to_csv(f'../data/press_releases/{new_file_name}', index=False)
