==> code/06-data-cleaning.ipynb <==
## Imports

import pandas as pd
import numpy as np

from tqdm import tqdm

import datetime as dt

import os

files = [f for f in os.listdir('../data/press_releases/') if 'csv' in f]
files

## Finding the dates for each file and assigning the label

### Apple

# find the dates
file = [f for f in files if 'apple' in f]
file
df = pd.read_csv(f'../data/press_releases/{file[0]}')

for row in range(len(df)):
    link_list = [i.strip() for i in df.loc[row,'full_link'].split('/') if i.strip() != '']
    df.loc[row,'date'] = '-'.join(link_list[3:5])

df['full_text'] = df['title'] + df['body']

df.date = df.date.map(lambda x: dt.datetime.strptime(x,'%Y-%m'))
df['year'] = df.date.map(lambda x: x.year)

df.head()

# assign the label

df['label'] = df.year.map(lambda x: 'after' if x >= 2020 else 'before')

# save it back

df.to_csv(f'../data/press_releases/{file[0]}', index = False)

### CVS Health

file = [f for f in files if 'cvs' in f]
df = pd.read_csv(f'../data/press_releases/{file[0]}')

years = [str(i) for i in range(2019,2022)]

for row in range(len(df)):
    para_list = [i.strip() for i in df.loc[row,'body'].split('\n') if i.strip() != '']
    df.loc[row,'date'] = '-'.join(para_list[3:6])
    
    for year in years:
        if (year in para_list[324]) & (len(para_list[324])<30):
            df.loc[row,'date'] = ','.join(para_list[324].split(',')[-2:]).strip()
            break
        
        elif (year in para_list[334]) & (len(para_list[334])<30):
            df.loc[row,'date'] = ','.join(para_list[334].split(',')[-2:]).strip()
            break
        
        else:
            df.loc[row,'date'] = None

df = df.dropna().reset_index(drop = True)

df['full_text'] = df.title + df.body


df.date = df.date.map(lambda x: dt.datetime.strptime(x,'%B %d, %Y'))
df['year'] = df.date.map(lambda x: x.year)

df.head()

# assign the label

df['label'] = df.year.map(lambda x: 'after' if x >= 2020 else 'before')

# save it back

df.to_csv(f'../data/press_releases/{file[0]}', index = False)

### Exxon Mobil

file = [f for f in files if 'exxon' in f]
df = pd.read_csv(f'../data/press_releases/{file[0]}')

# test = [i.strip() for i in df.loc[1,'body'].split('\n') if i.strip() != '']

# [(i,j) for i,j in enumerate(test)]

for row in range(len(df)):

    para_list = [i.strip() for i in df.loc[row,'body'].split('\n') if i.strip() != '']

    for year in years:
        
        if (year in para_list[78]) & (len(para_list[78])<20):
            df.loc[row,'date'] = ','.join(para_list[78].split(',')[-2:]).strip()
            break
        
        elif (year in para_list[75]) & (len(para_list[75])<20):
            df.loc[row,'date'] = ','.join(para_list[75].split(',')[-2:]).strip()
            break
        
        else:
            df.loc[row,'date'] = None

            
df['full_text'] = df.title + df.body


months = {
    'Jan.':'January',
    'Feb.':'February',
    'Mar.':'March',
    'Apr.':'April',
    'Jun.':'June',
    'Jul.':'July',
    'Aug.':'August',
    'Sept.':'September',
    'Oct.':'October',
    'Nov.': 'November',
    'Dec.':'December'
}

for row in range(len(df)):
    for old,new in months.items():
        df.loc[row,'date']=df.loc[row,'date'].replace(old,new).replace('   •','')

        
df['date'] = df['date'].map(lambda x: dt.datetime.strptime(x,'%B %d, %Y'))
df['year'] = df['date'].map(lambda x: x.year)

df.head()

# assign the label

df['label'] = df.year.map(lambda x: 'after' if x >= 2020 else 'before')

# save it back

df.to_csv(f'../data/press_releases/{file[0]}', index = False)

### Walmart

file = [f for f in files if 'walmart' in f]
df = pd.read_csv(f'../data/press_releases/{file[0]}')

for row in range(len(df)):
    para_list = [i.strip() for i in df.loc[row,'full_link'].split('/') if i.strip() != '']
    df.loc[row,'date'] = '-'.join(para_list[3:6])

df['full_text'] = df.title + df.body

df['date'] = df['date'].map(lambda x: dt.datetime.strptime(x,'%Y-%m-%d'))
df['year'] = df['date'].map(lambda x: x.year)

df.head()

# assign the label

df['label'] = df.year.map(lambda x: 'after' if x >= 2020 else 'before')

# save it back

df.to_csv(f'../data/press_releases/{file[0]}', index = False)

### Amazon

file = [f for f in files if 'amazon' in f]
df = pd.read_csv(f'../data/press_releases/{file[0]}')

df['full_text'] = df['title'] + df['body']

df.head()

# assign the label

df['label'] = df.year.map(lambda x: 'after' if x >= 2020 else 'before')

# save it back

df.to_csv(f'../data/press_releases/{file[0]}', index = False)

## Removing potential leak words

leak_words = [
    '2021', '2020', '2019', 'Covid-19', 'Covid', 'COVID-19', 'COVID',
    'Coronavirus', 'coronavirus', 'pandemic', 'Amazon', 'Apple', 'Walmart', 
    'CVS Health', 'CVS', 'Exxon Mobil', 'Exxon', 'menu','newsroom',
    'amazon.com', 'blog', 'press', '\n', '\r', '\t'
]

for file in tqdm(files):
    df = pd.read_csv(f'../data/press_releases/{file}')
    
    for word in leak_words:
        df.full_text = df.full_text.map(lambda x: x.replace(word,''))
    
    df.to_csv(f'../data/press_releases/{file}', index = False)
